Resolved the issue with telescope loading fzf
There are two different “fzf things” in play here:

fzf CLI tool (pkg install fzf)This is the standalone fuzzy finder you
can use in the shell.

Telescope itself does not use this binary. Instead it uses
telescope-fzf-native.nvim plugin This is a C extension that Telescope
compiles with make.

The alternative is to nstall telescope-fzf-writer.nvim.  So the
configuration was adjusted so it falls back gracefully — i.e. tries to
load telescope-fzf-native if compiled, but otherwise uses fzf_writer as
a pure Lua fallback